refactor(interfaces): move KvDtypeKind markers to ferrum-interfaces (Dim 5 cleanup)
The 5-dim architecture splits naturally:
- Dim 4 trait `Backend<B>` MUST live in ferrum-kernels — it has
  `type Buffer; type Context;` that need cudarc/metal types.
- Dim 5 markers (`KvDtypeKind`, `KvFp16` / `KvBf16` / `KvInt8` /
  `KvFp8`) are pure Rust types with no GPU deps. They belong in
  the contract crate (ferrum-interfaces), not in the kernel crate.

This PR moves just the markers. The capability trait
`BackendKvDtype<K>: BackendPagedKv` stays in ferrum-kernels because
its supertrait pulls in Backend.

Changes:
- New module `ferrum-interfaces/src/kv_dtype.rs` with KvDtypeKind +
  4 marker structs.
- Re-exported from ferrum-interfaces lib.rs alongside the other
  trait contracts.
- Original definitions in `ferrum-kernels/src/backend/traits.rs` replaced
  with `pub use ferrum_interfaces::kv_dtype::{...}` so existing
  `crate::backend::KvFp16` paths still resolve unchanged.
- ferrum-kernels gains a Cargo.toml dep on ferrum-interfaces (it
  already depended on ferrum-types).

Acceptance:
  cargo check --workspace --all-targets        clean
  cargo test --workspace --features metal --lib   clean (118 ferrum-engine)
