+++ crates/ferrum-kernels/src/backend/traits.rs
@@ -2153,42 +2153,11 @@ impl<T> MoeLlmBackend for T where T: QuantLlmBackend + BackendMoeFused {}
 // Future PR: implement BackendKvDtype<KvInt8> on CUDA + a new model
 // type-parameter `K: KvDtypeKind` to wire it through.
 
-/// Marker trait + metadata for a KV cache element type.
-pub trait KvDtypeKind: Send + Sync + 'static {
-    /// Stable name for logging / debug (e.g. "fp16", "int8").
-    const NAME: &'static str;
-    /// Bytes per element on disk + in cache memory.
-    const BYTES_PER_ELEM: usize;
-}
-
-/// FP16 KV cache (the existing default on CUDA + Metal).
-pub struct KvFp16;
-impl KvDtypeKind for KvFp16 {
-    const NAME: &'static str = "fp16";
-    const BYTES_PER_ELEM: usize = 2;
-}
-
-/// BF16 KV cache (drop-in replacement for FP16 on Ampere+ / Apple Silicon).
-pub struct KvBf16;
-impl KvDtypeKind for KvBf16 {
-    const NAME: &'static str = "bf16";
-    const BYTES_PER_ELEM: usize = 2;
-}
-
-/// INT8 KV cache — half the memory of FP16 with per-token / per-channel
-/// scale factors. CUDA path planned via vLLM's quant_kv kernels.
-pub struct KvInt8;
-impl KvDtypeKind for KvInt8 {
-    const NAME: &'static str = "int8";
-    const BYTES_PER_ELEM: usize = 1;
-}
-
-/// FP8 KV cache — E4M3 by default. Hopper+ on CUDA, future on Metal.
-pub struct KvFp8;
-impl KvDtypeKind for KvFp8 {
-    const NAME: &'static str = "fp8";
-    const BYTES_PER_ELEM: usize = 1;
-}
+// `KvDtypeKind` + `KvFp16` / `KvBf16` / `KvInt8` / `KvFp8` markers moved
+// to `ferrum_interfaces::kv_dtype` (no GPU deps, so the right place is
+// the contract crate). Re-exported here so existing callers keep
+// compiling against `crate::backend::KvFp16` etc.
+pub use ferrum_interfaces::kv_dtype::{KvBf16, KvDtypeKind, KvFp16, KvFp8, KvInt8};
 
 /// Capability-trait for backends that can store + read a KV cache of
 /// type `K`. Methods come in a follow-up PR; today this is a marker
